+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - --verbose argument
+- --traceback argument
 
 ## [0.1.1] - 2021-02-07